requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Configuration
pyyaml==6.0.1

# Async & Concurrency
aiohttp==3.9.1

# WebSocket
websockets==12.0

# Data Processing
numpy==1.26.2
pandas==2.1.3

# Alerting
python-telegram-bot==20.7
aiosmtplib==3.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
watchdog==3.0.0

# Monitoring
psutil==5.9.6
scipy==1.11.4
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
